add build tags for runtime choice
